Add --outfile option support

+++ runtest.sh
@@ -5,8 +5,8 @@
 ## (in name BASENAME[_SUBNAME].test) it may define
 ##    - ETS-configuration file (set by "config" variable. default: *.config)
 ##    - Template file (set by "template" variable. default: *.template)
-##    - *setup() shell function (default empty)
-##    - *teardown() function (default is to diff against BASENAME.expected)
+##    - setup() shell function (default empty)
+##    - teardown() function (default is to diff against BASENAME.expected)
 ##
 ## If a testcase-configuration is not exist or omits any of above definitions,
 ## the default file/function shall be used.
@@ -19,8 +19,8 @@
 ##    4. "foo.config" is searched.
 ##
 
-TEST_CASES="basic undef template_name outfile_name"
-TEST_CASES_FAIL="basic__t_opt basic__o_opt undef__fail"
+TEST_CASES="basic undef template_name outfile_name outfile_name__O_opt"
+TEST_CASES_FAIL="basic__t_opt basic__o_opt undef__fail outfile_name__O_and_o_opt"
 
 PYTHON="python2.5"
 TESTDIR="./test"
@@ -92,9 +92,10 @@ runtest ()
         ## Execute
         ##
         cmdline="$PYTHON ./ets.py ${option} ${config} ${template}"
-        echo $cmdline
+        echo "config: $cmdline"
         $cmdline 1> $stdout 2> $stderr
         command_ret=$?
+        echo "command_ret: $?"
 
         echo "--- stdout ----"
         cat $stdout
@@ -135,6 +136,18 @@ runtest ()
     return $?
 }
 
+##############################################################################
+case $1 in
+    '')
+        : ;;
+    -k|--keep)
+        opt_keep=yes;;
+    *)
+        echo "runtest.sh: ERROR: Unknown option: $1"
+        exit 1
+        ;;
+esac
+
 count_total=0
 count_fail=0
 
@@ -147,7 +160,7 @@ do
     if runtest $test > test-$test.log
     then
         echo OK
-        rm -f test-$test.log
+        [ "$opt_keep" = yes ] || rm -f test-$test.log
     else
         echo FAIL
         let count_fail++
@@ -161,7 +174,7 @@ do
     if runtest --fail $test > test-$test.log
     then
         echo OK
-        rm -f test-$test.log
+        [ "$opt_keep" = yes ] || rm -f test-$test.log
     else
         echo FAIL
         let count_fail++
